Ultimo commit. Trabajando sobre GAF

+++ IA - GeneticAlgorithm -TP/IA_TP.Model/Catalogue.cs
@@ -2,8 +2,8 @@
 {
     public class Catalogue
     {
-        public decimal TV_Price { get; set; }
-        public decimal Internet_Price { get; set; }
-        public decimal Phone_Price { get; set; }
+        public double TV_Price { get; set; }
+        public double Internet_Price { get; set; }
+        public double Phone_Price { get; set; }
     }
 }
